Introduce test infra for the Gradle plugin
This commit introduces functional tests for the Gradle plugin.
For now, it's limited to a single test but it will be expanded for
more use cases. Functional tests are using samples found in the
`src/samples` directory.

There are 2 test tasks which can be executed, depending on whether
you want full test coverage or not:

- `functionalTest` will run the test suite on the current version
of Gradle only
- `fullFunctionalTest` will do the same, but on all supported Gradle
versions

To make this possible, the build has been reorganized to leverage
Gradle composite builds, instead of using the local Maven repository.
Previously to this, it was necessary to cd into `junit-platform-native`,
publish to maven Local, then go to the Gradle plugin, build it, and
test with a local project.

This is no longer required, as now the execution of functional tests
will _automatically_ trigger the build of the native JUnit project,
publish the library to a _test_ repository, which in turn is used by
the functional tests (via repository injection in an init script).

This makes it impossible to make mistakes, or have the build fragile
if you don't build the `junit-platform-native` first: the dependency
is now explicit.