Simplify task configuration
Previously to this commit, there were 2 extensions of 2 different types
to configure the native extension, depending on whether it was for tests
or not.

It wasn't necessary, as the configuration of the native image task is
the same in both cases, it's simply the values of the extensions which
differ slightly. Therefore, this commit simplifies everything by using
a _single_ extension type.

Similarly, there were 2 "build image" task types, when in practice the
building mechanism is the same in both cases. Therefore, the 2 tasks
were replaced with a single one, `BuildNativeImageTask`, which is setup
appropriately in both cases.

Last, for the same reason, the 2 "run" task types have been replaced
with a single version.

Last but not least, configuration of the tasks is now properly done
by the plugin, instead of being mixed between the plugin configuration
and the task creation. There were ordering issues due by the fact that
the configuration of the task was mutating its own state.

Instead, tasks are now properly isolated, without any knowledge of
their configuration, and the plugin is responsible for injecting
the configuration of the tasks.

This makes it possible, for example, to derive the name of the test
image from the name of the main image.

It's worth noting that this commit didn't fully refactor the configuration
phase, because there's still fishy configuration of the 'agent' side
of things. This will be done in a subsequent commit.