Rework how the agent is injected
This commit reworks how the agent is injected on the `run` and `test`
tasks. There is now a task which role is to copy the filter json file,
but more importantly the value of the `agent` property is not read
eagerly.

A new test has been added to make sure that the agent is properly
injected and that the configuration files are generated. This doesn't
work properly for GraalVM versions <21.1.

The "persistConfig" property is temporarily _unused_, we have to
decide what to do with it.