Remove persistConfig
There were several issues with this flag:

1. it persists outputs in the (re)sources directories, which means
that potentially, during a build, a task can change the contents
of the (re)sources. Therefore, there's a potential ordering and
non reproducibility issue based on when a task sees the resources.

2. it arbitrarily chooses a source set to persist data, but in
practice this shouldn't live in a source set, since the result
of the analysis, and actually the resources needed, may be split
accross different source sets (think Java vs Kotlin sources)

Therefore we agreed on removing it. A user can still access the
result of the agent analysis via the build directory, and copy
them manually if needed, or they can implement their own task,
with their own risks taking, if they really want to do this.