Replace agent-filter.json with post-processing
This commit replaces the adhoc `agent-filter.json` file with a post-processing
step. Basically, instead of passing an access filter file to the agent, we now
post-process the generated JSON files to remove entries that we don't need.

There are several advantages doing this:

1. user can rewire the image build task to use different files
2. filering can be made more configurable in the future
3. the plugin is now compatible with older GraalVM releases

The current way the files are filtered, though, is very rough. Because there's
no official JSON schemas for the generated files, it's an adhoc, deep inspection
algorithm which filters known patterns.