Introduce task to generate a resource-config.json file
This commit introduces a new task which will generate a resource-config.json file.
There are 3 ways this task can be configured:

- with explicit configuration of includes, excludes and bundles
- with inference of includes and excludes
- a combination of both

By default, the task will do nothing, meaning that the generated file
will be empty. To declare resources to include, the user needs to configure
the `resources` block of the extension:

```
nativeBuild {
   resources {
      bundles.add('my.bundle')
      includedPatterns.add('some resource regex')
   }
}
```

Alternatively, in inference mode, Gradle will try to infer the resources
to include from what is found on classpath:

```
nativeBuild {
    resources {
        inference {
           enabled = true
        }
    }
}
```

Inference is disabled by default, and will also be limited to project
artifacts (including transitive project artifacts in a multi-project
build). It might be interesting to include scanning of _all_ transitive
dependencies (including external dependencies) when performing inference,
at the cost of probably more noise in the generated file:

```
nativeBuild {
    resources {
        inference {
            enabled = true
            restrictToProjectDependencies = false
        }
    }
}
```

It is also possible to control the output of the inferred resources,
for example to exclude some resources:

```
nativeBuild {
    resources {
        inference {
             enabled = false
             restrictToProjectDependencies = false
             inferenceExclusionPatterns.add('^do/not/include/this/resource$')
        }
    }
}
```

Fixes #75