First step in consolidating the build
This commit introduces a parent build which is responsible for coordinating the
build of every subproject, independently of what they are using (Maven or Gradle).
It's a Gradle composite build, which takes the responsibility of publishing
subprojects artifacts into a "common" local repository, which can be used to
verify by hand what is produced by the builds easily.

On the root project, you can now run
    `./gradlew publishAllPublicationsToCommonRepository`

which will trigger the dependent builds and publish all artifacts to:

    ./build/common-repo