Introduce an utils project
Before this commit, both the Maven and Gradle plugins had an identical
VersionInfo class. Instead of this, we now introduce a `utils` project
which generates this class statically at build time.

This project will be used to share common code between the Gradle and
Maven plugins, even if today it only contains a single class.