Introduce functional testing of the Maven plugin
Testing the Maven plugin was complicated. Not because we build it with Gradle,
but because running any build with Maven can accidentally pass because of the
presence of stale files in the local Maven repository.

In this case, we avoid the problem by completely isolating the Maven plugin
tests by running them in containers, using test containers. Because spawning
an image and re-downloading all dependencies can be pretty slow, we proceed
in multiple steps:

1. we build the plugin, and all its dependencies, and publish the result in
the local "common repository"
2. we create a docker image which base is the GraalVM image for the version
we want to test
3. we inject a "bootstrapping" Maven project in that image, which is going
to be responsible for triggering the download of the dependencies required
at test execution time, so that all dependencies are present in the image
when we want to run the tests
4. because the project needs the plugin to be built, we also inject a fake
Maven repository, the one we built at step 1, into the image
5. we use a custom settings file for Maven so that the repository is
injected in the build
6. now we can build the Docker image
7. integration tests spawn the test image and perform some verifications

It's worth noting that at this stage, there's a *single* functional test
for Maven. This test is very similar to the one done in Gradle, so, instead
of providing examples in the examples directory, we will refactor the
build so that the samples used for integration testing are also samples
that can be shown to users.