Add a release zip task
This commit adds a task, `releaseZip` which will build all artifacts
and create a zip in `build/distributions`. This task makes sure that
the repository is pruned before building, so that it only contains
what we want to release.

The task will _not_ update the version, which needs to be done
separately if needed.