Introduce a top-level `graal` DSL container
This is a follow-up to the idiomatic Gradle conversion. Before this commit,
there were 2 extensions configured at the top level, with the same type,
to configure the main native image and the test native image. Those
extensions were respectively named `nativeBuild` and `nativeTest`. In
addition, there were 2 tasks with similar names, which had slightly
different semantics: `nativeBuild` would build the main image, but
`nativeTest` would _run_ the tests. The task which would build the native
test image was called `nativeTestBuild`.

There were therefore a few issues to fix:

- the confusion between the extension names and the task names, making it
unclear whether you were configuring one or the other
- the fact that multiple extensions of the same type were created, which is
considered an anti-pattern in Gradle and makes it complicated to get the
expected extensions (See https://github.com/gradle/gradle/issues/17622)
- it was complicated to add support for more test images, as hinted by
the follow up issue #77

So this commit changes the configuration to use an idiomatic container,
named `graal` which can contain _many_ images. For now it contains only
two:

```
graal {
    main {
        // .. the main native image
    }
    test {
        // .. the test image
    }
}
```

In addition, it renames the confusing task names to something closer to the
actual semantics of other tasks in Gradle:

- `nativeBuild` -> `nativeAssemble` (`build` does much more)
- `nativeTestBuild` -> `nativeTestAssemble`