Introduce documentation project
This commit introduces documentation built with Asciidoctor.
Instead of using README files sparsed into the various subproject
directories, this introduces a first-class citizen documentation
project which contains documentation for both the Maven and Gradle
plugins.

In addition, code is ready for publication of this documentation
under GitHub pages, which means that once this gets merged, we
will be able to setup an initial GitHub pages branch, and a workflow
which automatically publishes updated documentation.

Each version will be published into its own subdirectory, meaning
that we will be able to point users to documentation specific to
their usage of the plugins.

Because this is the first commit, the README haven't been updated
yet: they will need to be changed to be _developer documentation_
only and link to the user docs, once the first GitHub pages are
published.