Use an extension to configure the Maven build
This commit introduces a Maven extension which is used to tweak
the surefire plugin configuration (in particular to enable the
test listener). This also has the advantage of letting us get
rid of the junit-platform-native dependency.

Now users will have to enable the extension when configuring
the plugin, but will no longer need to add extra dependencies.