Introduce configurations for configuring the classpath
Previously to this commit, the classpath used by native image was
hardcoded to include the project artifact and its runtime classpath.
With this change, the plugin now introduces a couple of configurations:

- `nativeImageCompileOnly` which _can_ be used to declare dependencies
(or artifacts) which are required at image build time, but not by the
main application (JVM mode)
- `nativeImageClasspath`, which extends from the same configurations
as the `runtimeClasspath` configuration, but also from the new
`nativeImageCompileOnly` configuration. This is the configuration which
is now used to configure the plugin classpath.

This will allow other plugins to contribute more artifacts or dependencies
at image build time without having to tweak manually the image classpath,
which could before imply loosing some user configuration code.

This also makes it cleaner to contribute GraalVM native image features,
without leaking into the main application runtime classpath.

Eventually, the same mirror configurations are defines for the test
image building.

Closes #129