Add support for agent in Maven plugin
This commit adds intergration with the GraalVM agent to the Maven
plugin. Enabling agent support requires setting the `-Dagent=true`
flag on the command line. BBecause of technical limitations in Maven,
the process is not as smooth as with Gradle.

In particular, while running tests with the agent is straightforward,
doing the same with the application itself requires configuring a
distinct Mojo execution of the `maven-exec-plugin`.

Fixes #13