Add support for multiple test images
This commit introduces an API to register more test suites as
native images. This is useful whenever a project defines multiple
test tasks and/or test source sets. Previously the plugin would
only create a test image for the main test suite, now it's possible
to register more test suites to be executed natively.

Closes #77