Fail native test build for failed tests and containers only
Prior to this commit, a native test build failed if there were any
failed tests or aborted tests within the native image. However, aborted
tests should not fail the build. In addition, prior to this commit a
failed container did not fail the native test build. A container (i.e.,
typically a test class) can fail if a class-level callback or extension
fails -- for example, if an exception is thrown from a JUnit Jupiter
@BeforeAll method.

This commit addresses these two issues by making the exit code for
the NativeImageJUnitLauncher based on the total failure count which
includes the test failure count and the container failure count.

An aborted test has been added to BasicTests. An @BeforeAll method
has also been added to BasicTests with a code block that throws an
exception when running within a native image. This code block has been
commented out to ensure that the build passes after the change in this
commit, but if you reinstate that code block without the change the
build would pass even though the test class failed.

Fixes #153