Relocate TestIdentifier build-time init to PlatformConfigProvider
Prior to this commit, the build-time initialization for the JUnit
Platform TestIdentifier type was registered directly in the
JUnitPlatformFeature. This works fine; however the configuration
for the JUnit Platform, Jupiter, and Vintage projects each has its own
dedicated PluginConfigProvider implementation.

For the sake of consistency, this commit therefore relocates the
TestIdentifier build-time init registration to the PlatformConfigProvider.

Fixes #150