Support custom agent options in Maven plugin
Prior to this commit, users could not supply custom options to the JVM
agent. In addition, the `experimental-class-loader-support` option was
always supplied, which may not be desired by all users.

This commit introduces support for configuring the agent by allowing
users to specify multiple sets of named `agentOptions` in the
`configuration` block for the native-maven-plugin.

The following example configures two sets of agent options, one for
execution of the application (exec) and one for tests (test).

<configuration>
  <agentOptions name="exec">
    <agentOption>experimental-class-loader-support</agentOption>
  </agentOptions>
  <agentOptions name="test">
    <agentOption>experimental-class-loader-support</agentOption>
    <agentOption>access-filter-file=${basedir}/src/test/resources/access-filter.json</agentOption>
  </agentOptions>
</configuration>

In addition, the `experimental-class-loader-support` option is no
longer configured automatically by the native-maven-plugin.

See #162