Make it possible to configure agent options
Before this commit, the `experimental-class-loader-support` was
systematically added to the agent options and it wasn't possible
to configure the options passed to the agent.

Now, each image can configure agent options independently. As
part of this work, cleanup was done in the way agent configuration
was injected. In particular, each native image now has to explicitly
state what _task_ should be instrumented when the agent is enabled.
The plugin automatically configures the `main` image to instrument
the `run` task, while the `test` image instruments the `test` task.

The agent options are configured via a new `agentOptions` block.
The old "agent.set(true)` method is delegating to the `agentOptions`
block.

See #162