Allow agent to be enabled within the Maven POM
Prior to this commit, the GraalVM native agent was disabled by default
and could only be enabled by supply `-Dagent=true` to the Maven command
line.

With this commit, the agent is still disabled by default, but it can
now be enabled within the Maven POM in the configuration for the native
Maven plugin (see <agent> -> <enabled> in the example below).

If the agent is enabled in the POM, it can be disabled via the command
line by supplying `-Dagent=false`.

In addition, this commit introduces a new <agent> node in the
configuration that serves as a container for all agent related
settings. Consequently, the agent options are now configured via
<agent> -> <options> -> <option>.

The following example demonstrates the supported configuration.

<agent>
    <enabled>true</enabled>
    <options name="main">
        <option>experimental-class-loader-support</option>
    </options>
    <options name="test">
        <option>experimental-class-loader-support</option>
        <option>access-filter-file=${basedir}/src/test/resources/access-filter.json</option>
    </options>
</agent>

See #162