Support shared, unnamed agent options in the Maven plugin
Prior to this commit, <options> elements were required to declare a
unique name attribute and could only be activated via the
`-DagentOptions=<NAME>` command line argument.

This commit introduces support for an unnamed <options> element that is
always active whenever the agent is used. This unnamed set of options
is intended to be used for common options that are shared across all
agent executions.

The following demonstrates the use of this new feature.

<agent>
  <enabled>true</enabled>
  <options><!-- unnamed, shared options -->
    <option>experimental-class-loader-support</option>
  </options>
  <options name="main"><!-- main options -->
    <option>config-write-period-secs=30</option>
    <option>config-write-initial-delay-secs=5</option>
  </options>
  <options name="test"><!-- test options -->
    <option>access-filter-file=${basedir}/src/test/resources/access-filter.json</option>
  </options>
</agent>

See #162