Support main/test agent options by default in the Maven plugin
Prior to this commit, the Maven plugin supported unnamed and named sets
of <options> for the agent, but there were no predefined sets of
options such as "main" and "test" which the Gradle plugin supports.

This commit aligns the Maven support for agent options with the Gradle
support by introducing explicit support for "main" and "test" agent
execution contexts.

Whenever the agent is executed with the application, the "main" set of
agent options will be automatically activated.

Similarly, whenever the agent is executed with the tests, the "test"
set of agent options will be automatically activated.

See #162