Let resources autodetection work if file contains native-image config
Previously to this commit, resources autodetection was disabled if a
classpath entry (directory or jar file) contained a `META-INF/native-image/xxx/resource-config.json`
file. This was done assuming that if that file was present, then it
_should_ contain all the appropriate entries already. However, this
is not the case in Micronaut where an annotation processor is
generating such a file, but doesn't include _all_ the resources of the
application.

As a workaround, it's now possible to ask the scanner to detect
resources even if that file is present.

See https://github.com/micronaut-projects/micronaut-gradle-plugin/issues/327