Rework how disabling of toolchain detection works
Disabling the toolchain detection by tasks wasn't sufficient,
because Gradle would use the nested `BaseImageOptions` as an input,
and even if the particular `JavaLauchner` input wasn't used if that
flag was disabled, it would still be used internally by Gradle to
snapshot inputs.

As a consequence, the fix only worked if Gradle actually detected
a GraalVM installation somewhere, even if it didn't use it!

To fix this, this commit introduces a top-level DSL method called
`disableToolchainDetection()` which effectively sets the toolchain
to `null` by convention. By making the toolchain optional and null
now Gradle wouldn't complain anymore.

Fixes #183