Fix race condition if tests are executed in parallel
This commit makes use of the `native-image-configure` tool to merge
the results of agent files. Before this commit, several process could
be writing to the same output directory. Now, each process will use
a separate directory and the tool is used to merge their outputs.

Because the `native-image-configure` tool might not be present,
it's also going to invoke the command which generates it.

See #194