Do not use separate agent directories when agent is invoked on main
This is a workaround for goal execution ordering problems with Maven.
For tests, we can add a synthetic goal which is executed in the same
phase and will merge the agent files generated by the test execution.

Unfortunately, this approach doesn't work if we want to instrument
the main execution, which is sometimes the case. This is not doable
because we won't have a phase to hook into and a goal to execute in
order to merge the files. Even if we synthetize a goal and try to
add it to the session execution, this would cause ordering issues
because we don't know in which order the plugins are defined in the
build file. In addition, this would simply not work for multi-module
builds.

As a workaround, we simply disable the split agent output directories
in case we instrument the main execution. This will hopefully work
for most invocations.