Use argument file when invoking `native-image`
With this, it is no longer necessary to build a fat jar (or shaded jar)
under Windows in case of long classpath issues. Using the argument
file is enabled by default. If the user uses a version older than 21.3,
then they can use the fat jar support (and disable arg-file usage).

Fixes #203