Add integration testing for native config repository
This commit introduces testing of the native configuration repository
integration for Gradle. Instead of using the "real" configuration
repository, which would require writing a test library which uses
"unsupported" GraalVM features, the tests use a test fixture, a
library which makes use of reflection. This library is injected
into the tests and we create a corresponding test repository.