Add DirectoryConfiguration.copy method
Update `DirectoryConfiguration` with a new `copy` method that
can be used to copy matadata files into `META-INF/native-image`
directories.

The `DirectoryConfiguration` class now has group, artifact and
version information so that it can write files into the
appropriate folder. The `copy` method will also write a
`reachability-metadata.properties` file that indicates if the
'override' property was set on the metadata.