Fix functional tests for MacOS users (#321)
Update `AbstractGraalVMMavenFunctionalTest` web server setup to
use real paths. Prior to this commit, when running the functional
tests on MacOS the web server would not serve the `target/repo.zip`
file. This is due to the fact that the temp folder on MacOS can
be a symlink which by default Jetty will not serve. Switching
the resource base to the real folder fixes the problem.