Download the metadata repository as a publishable artifact
This commit introduces a task which is responsible for downloading the
current metadata repository version and bundle it as a publishable artifact.
This means that the ZIP file of the metadata repository is going to be
downloaded from the GraalVM metadata repository, and published as a
_classified_ artifact on the `graalvm-reachability-metadata` module.

It is possible to check the result by calling:

    ./gradlew graalvm-reachability-metadata:pAPTCR

and checking the contents of the `build/common-repo` directory. A file
named `graalvm-reachability-metadata-<nbt version>-repository.zip` file
should be visible.

This will make it possible to consume the metadata repository as a Maven
artifact.