Make the Gradle plugin use the metadata repository from Maven
This commit makes it so that if the user doesn't specify any URL for the
repository, nor any specific version, then we would use the repository
artifact which is bundled with the NBT release on Maven Central (or
any repository configured in the build). This makes it possible to
use the Gradle native build tools exclusively with Maven repositories.