Include buildtool support for bundling reachability metadata into jar files (#322)
* Add DirectoryConfiguration.copy method

Update `DirectoryConfiguration` with a new `copy` method that
can be used to copy matadata files into `META-INF/native-image`
directories.

The `DirectoryConfiguration` class now has group, artifact and
version information so that it can write files into the
appropriate folder. The `copy` method will also write a
`reachability-metadata.properties` file that indicates if the
'override' property was set on the metadata.

* Add `add-metadata-hints` goal to Maven Plugin

Update the maven plugin with a new `add-metadata-hints` goal that can
be used to bundle hints obtained from the metadata repository into
the jar.

* Add `ReachabilityMetadataCopyTask` to Gradle Plugin

Update the gradle plugin with a new `ReachabilityMetadataCopyTask`
class that can be used to copy hints obtained from the metadata
repository.

* Refactor common code in NativeImagePlugin

Refactor the `configureJvmReachabilityConfigurationDirectories`
and `configureJvmReachabilityExcludeConfigArgs` to use a common
method since they share the same logic.