Ship the GraalVM metadata repository as an artifact alongside NBT (#331)
* Bump the metadata repository version

This also moves the constant to the `libs.versions.toml` file so that it
is both easier to upgrade *and* available in build scripts without having
to parse source files. This is going to be useful for repackaging the
repository.

* Download the metadata repository as a publishable artifact

This commit introduces a task which is responsible for downloading the
current metadata repository version and bundle it as a publishable artifact.
This means that the ZIP file of the metadata repository is going to be
downloaded from the GraalVM metadata repository, and published as a
_classified_ artifact on the `graalvm-reachability-metadata` module.

It is possible to check the result by calling:

    ./gradlew graalvm-reachability-metadata:pAPTCR

and checking the contents of the `build/common-repo` directory. A file
named `graalvm-reachability-metadata-<nbt version>-repository.zip` file
should be visible.

This will make it possible to consume the metadata repository as a Maven
artifact.

* Make the Gradle plugin use the metadata repository from Maven

This commit makes it so that if the user doesn't specify any URL for the
repository, nor any specific version, then we would use the repository
artifact which is bundled with the NBT release on Maven Central (or
any repository configured in the build). This makes it possible to
use the Gradle native build tools exclusively with Maven repositories.

* Make the Maven plugin use the metadata repository from Maven Central

Similarly to the Gradle plugin, this commit makes it so that if the user
doesn't specify any URL for the repository, nor any specific version,
then we would use the repository artifact which is bundled with the NBT
release on Maven Central.

* Add release notes

* Add mention about the new metadata collection task