Fix compile task out of date when runtime args change (#372)
* Fix compile task out of date when runtime args change

The `NativeImageOptions` domain object is used in the DSL to configure
both compile and runtime arguments for native images: the compile
options are used to invoke `native-image` and build a binary, while
the runtime args are used to pass extra arguments when executing the
native image.

The full `NativeImageOptions` domain object was passed to the
native image compilation task as a `@Nested` input, which meant in
practice that if the _runtime_ arguments changed, then we would
rebuild the image although it isn't necessary.

This commit fixes the problem in a binary compatible way, by
introducing a couple of interfaces: one for the compile options,
the other for the runtime options. It makes the "options" internal
on the compile task, and uses a delegated view of compile options
for input snapshotting.

Fixes #371

* Remove pre-Gradle 7 tests

With the removal of the Java 11 variant of GraalVM, we're not in capacity
of testing this on pre-Gradle 7 since the spec won't let us specify that
we need a GraalVM vendor.