Fix deadlock when collecting metadata
The reachability metadata service is not thread-safe, but it is
possible for it to be called from multiple tasks in a multi-project.
It this happens, then a deadlock may happen, or weird errors with
a property not set can be thrown at the user.

This fixes the problem by making sure only a single task can
access the service concurrently. Ideally, the service should be
reworked to allow concurrent access, but it's quite some work
since it involves locking access to the repository (typically
to download and unzip metadata).

Fixes #387