Add a new write-args-file goal
This PR introduces a goal that only writes the args file, so that other plugins further in the lifecycle can use it. In particular, the Micronaut Maven Plugin needs this goal.

The path to the args file is stored in the `MavenProject` properties, an approach that is used by other plugins such as the JaCoCo Maven Plugin.