Disable toolchain detection by default
With the changes coming in GraalVM regarding versioning, in particular
the fact that the GraalVM version becomes internal and that the
vendor string is changing, toolchain detection will clearly be broken
for all users.

This commit therefore disables toolchain detection by default, and
makes sure to document properly the new behavior.