Fix lookup of metadata repository
This commit fixes the URI which is constructed when only a version is set in the
metadata repository. Before this change, whatever version was used in the configuration,
we would ignore it, because we were testing against the wrong URI to figure out if the
user had configured something explicitly or not.

Now, we will prioritize the user URI, then construct a GitHub release URI from the
version which is configured, and eventually use Maven Central if that URI is the
same as the default one.

Fixes #424