Add support for `default-for` attribute
This new index.json attribute is design to allow more
flexible version matching to avoid breakage due to
switch on latest version for untested ones.

It should have a `default-for` key and a regexp as a
string value, for example "1\\.0\\..*" which should
match versions like "1.0.0", "1.0.1", "1.0.2", etc.

See oracle/graalvm-reachability-metadata#62
See oracle/graalvm-reachability-metadata#275