Add support for PGO instrumentation (#471)
* Add support for PGO instrumentation

This commit adds support for PGO instrumentation. This should be
enabled by adding the `--pgo-instrument` option to the Gradle
command line. When this is done, then the generated binary will
be compiled with PGO instrumentation enabled, and the binary
name will be suffixed with `-instrumented`.

It is possible to run the instrumented binary directly too,
in which case the profile files will be written in the same
directory as the binary.

* Add support for a PGO profiles directory

By convention, the directory is set to `src/pgo-profiles/<binary>`.
For example, for the `main` binary, the directory where to put
PGO profiles would be `src/pgo-profiles/main`. If that directory
is present _and that we're not instrumenting_, then the profile
will be used when compiling with native image.

It is possible to provide multiple profiles in a single directory.

* Remove GraalVM version from workflows

* Add documentation about PGO support

See #457

* Fix JUnit native test

* Make checkstyle happy

* Fix tests

* Temporarily(?) disable testing with config cache

As we're not compatible. Test `org.graalvm.buildtools.gradle.OfficialMetadataRepoFunctionalTest`
throws an incomprehensible error message, in all versions of Gradle I've tested:

```
Configuration cache state could not be cached: field `spec` of `org.gradle.api.internal.tasks.execution.SelfDescribingSpec` bean found in task `:compileJava` of type `org.gradle.api.tasks.compile.JavaCompile`: error writing value of type 'org.gradle.api.internal.tasks.compile.CompilerForkUtils$$Lambda$1235/0x00000008015b1c38'
> Unable to make field private final java.lang.Object[] java.lang.invoke.SerializedLambda.capturedArgs accessible: module java.base does not "opens java.lang.invoke" to unnamed module @3cc98b0c
```

This PR also rewrote some code which fixed other configuration
cache issues which arose _before_ reaching this one.

* Upgrade to JUnit 5.10.0

* Make checkstyle happy

* Fix test

* Restore configuration cache tests

* Update baseline versions for config cache